samples: knot: Handle GET_DATA request
This patch adds the feature to handle get gata requests from the cloud.
As there is no need to wait for an answer for GET_DATA responses, a
flag was added to inform that.

The behaviour of process_event() will have to consider that the flag
proxy.wait_resp should be set, as data messages due polling check
should wait for an DATA_RESP.

Signed-off-by: Joao Cordeiro <jvcc@cesar.org.br>