samples: knot: Reset flash by button timeout
This patch introduces the peripheral.c file to handle peripherals and
their related expected behaviour.

The peripherals should be initialized by proto.c and used by sm.c, just
as storage controller.

After the reset button is pressed after 2 seconds, a flag is set. The
state machine reads the flag and can decide if and how it should reset.

Signed-off-by: Joao Cordeiro <jvcc@cesar.org.br>