samples: knot: Split flash's storage by OT and NVS
Reduce number of pages used by OpenThread to store its settings to 2.
As OpenThread is always placed at the end of flash, this moves up the
bottom of OpenThread's flash area.

Changed NVS flash area offset to beginning of storage area so it is
evenly splited by OpenThread and NVS.

Signed-off-by: Joao Cordeiro <jvcc@cesar.org.br>