samples: knot: Store app credentials as a struct
Change storage api module to pack UUID, TOKEN and ID in a single
struct to ease handling and save storage ids. This is permited as
they are all stored and recovered in the same contexts.

To ease futher migrations to zephyr's Settings sub-system, the
variables 'uuid', 'token', and 'device_id' from sm.c where kept.
The struct is used only for storing and recovering values from nvs.

Signed-off-by: Joao Cordeiro <jvcc@cesar.org.br>